+++ packages/dev/sharedUiComponents/src/components/reactGraphSystem/GraphContainer.modules.scss
@@ -4,6 +4,5 @@
     font: 14px "acumin-pro";
     user-select: none;
     overflow: hidden;
-    // cursor: move;
     background-image: linear-gradient(to right, #4f4e4f 1px, transparent 1px), linear-gradient(to bottom, #4f4e4f 1px, transparent 1px);
 }
