--- packages/tools/vsm/src/components/types.ts
@@ -1,7 +1,0 @@
-export enum NodeTypes {
-    StateNode = "StateNode",
-    StartActionNode = "StartActionNode",
-    ActionNode = "ActionNode",
-    ReadyActionNode = "ReadyActionNode",
-    ValueNode = "ValueNode",
-}
